Add files via upload
The project activity is a 10 MW grid connected solar PV power project in District Surendernagar in the State of Gujarat, India. The promoter of the project activity is Waa Solar Limited. Waa Solar Limited has executed a contract agreement with Madhav Infra Projects Pvt. Ltd (MIPPL) to execute civil, mechanical and electrical works for the proposed project activity. The project is an operational activity with continuous reduction of GHG (a registered CDM Project with ID 6352).